Add files via upload
This is a version of the extrapolation without the interpolated values. No final forecasts or confidence intervals are included here yet since some results look a bit off. Otherwise, this is the setup I consider for the extrapolation, with the inclusion of the interpolated values too. If you guys agree with this setup.